Move 'ReactTypeOfMode' to shared, use in 'ReactDebugCurrentFiber'
**what is the change?:**
- Moves 'ReactTypeOfMode' to the 'shared' directory
- Uses 'ReactTypeOfMode' and 'ReactDebugCurrentFiber' to detect when we
are in strict mode, allowing updating the warnings in the reconciler.

**why make this change?:**
We are in the process of prefixing all warnings with 'Strict Mode'.

**test plan:**
Run unit tests.

**issue:**
Internal task https://our.intern.facebook.com/intern/tasks/?t=30481686
